requirements.txt
sentence_transformers
numpy<2.0
faiss-cpu
google-generativeai
python-dotenv
streamlit
stripe